Remove devshop_container on robo destroy, also ask before killing everything please?? :D

+++ RoboFile.php
@@ -379,8 +379,13 @@ class RoboFile extends \Robo\Tasks
    * Running with --force
    */
   public function destroy($opts = ['force' => 0]) {
-    $this->_exec('docker-compose kill');
-    $this->_exec('docker-compose rm -fv');
+    
+    if ($this->confirm("Destroy docker containers and volumes?")) {
+      $this->_exec('docker-compose kill');
+      $this->_exec('docker-compose rm -fv');
+      $this->_exec('docker kill devshop_container');
+      $this->_exec('docker rm -fv devshop_container');
+    }
     
     $version = self::DEVSHOP_LOCAL_VERSION;
     $uri = self::DEVSHOP_LOCAL_URI;
